@@ -20,7 +20,7 @@ import android.Manifest
 import android.app.AlertDialog
 import android.app.Dialog
 import android.os.Bundle
-import android.support.v4.app.DialogFragment
+import androidx.fragment.app.DialogFragment
 
 /**
  * Shows OK/Cancel confirmation dialog about camera permission.
@@ -29,7 +29,7 @@ class ConfirmationDialog : DialogFragment() {
 
   override fun onCreateDialog(savedInstanceState: Bundle?): Dialog =
     AlertDialog.Builder(activity)
-      .setMessage(R.string.request_permission)
+      .setMessage(R.string.tfe_pn_request_permission)
       .setPositiveButton(android.R.string.ok) { _, _ ->
         parentFragment!!.requestPermissions(
           arrayOf(Manifest.permission.CAMERA),
