@@ -17,9 +17,9 @@ package org.tensorflow.lite.examples.smartreply;
 
 import android.content.Context;
 import android.content.res.AssetFileDescriptor;
-import android.support.annotation.Keep;
-import android.support.annotation.WorkerThread;
+import androidx.annotation.Keep;
 import android.util.Log;
+import androidx.annotation.WorkerThread;
 import java.io.BufferedReader;
 import java.io.FileInputStream;
 import java.io.IOException;
