@@ -18,7 +18,7 @@ package org.tensorflow.lite.examples.textclassification;
 
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.widget.Button;
@@ -41,7 +41,7 @@ public class MainActivity extends AppCompatActivity {
   @Override
   protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
-    setContentView(R.layout.activity_main);
+    setContentView(R.layout.tfe_tc_activity_main);
     Log.v(TAG, "onCreate");
 
     client = new TextClassificationClient(getApplicationContext());
