@@ -13,9 +13,8 @@
 // limitations under the License.
 // =============================================================================
 
-import Foundation
+import UIKit
 import Accelerate
-
 
 extension CVPixelBuffer {
 
@@ -97,4 +96,48 @@ extension CVPixelBuffer {
 
     return thumbnailPixelBuffer
   }
+
+  static func buffer(from image: UIImage) -> CVPixelBuffer? {
+    let attrs = [
+      kCVPixelBufferCGImageCompatibilityKey: kCFBooleanTrue,
+      kCVPixelBufferCGBitmapContextCompatibilityKey: kCFBooleanTrue
+    ] as CFDictionary
+
+    var pixelBuffer: CVPixelBuffer?
+    let status = CVPixelBufferCreate(kCFAllocatorDefault,
+                                     Int(image.size.width),
+                                     Int(image.size.height),
+                                     kCVPixelFormatType_32BGRA,
+                                     attrs,
+                                     &pixelBuffer)
+
+    guard let buffer = pixelBuffer, status == kCVReturnSuccess else {
+      return nil
+    }
+
+    CVPixelBufferLockBaseAddress(buffer, [])
+    defer { CVPixelBufferUnlockBaseAddress(buffer, []) }
+    let pixelData = CVPixelBufferGetBaseAddress(buffer)
+
+    let rgbColorSpace = CGColorSpaceCreateDeviceRGB()
+    guard let context = CGContext(data: pixelData,
+                                  width: Int(image.size.width),
+                                  height: Int(image.size.height),
+                                  bitsPerComponent: 8,
+                                  bytesPerRow: CVPixelBufferGetBytesPerRow(buffer),
+                                  space: rgbColorSpace,
+                                  bitmapInfo: CGImageAlphaInfo.noneSkipLast.rawValue) else {
+      return nil
+    }
+
+    context.translateBy(x: 0, y: image.size.height)
+    context.scaleBy(x: 1.0, y: -1.0)
+
+    UIGraphicsPushContext(context)
+    image.draw(in: CGRect(x: 0, y: 0, width: image.size.width, height: image.size.height))
+    UIGraphicsPopContext()
+
+    return pixelBuffer
+  }
+
 }
