@@ -15,6 +15,7 @@
 import CoreImage
 import TensorFlowLite
 import UIKit
+import Accelerate
 
 /// A result from invoking the `Interpreter`.
 struct Result {
@@ -105,6 +106,7 @@ class ModelDataHandler {
 
   /// Performs image preprocessing, invokes the `Interpreter`, and processes the inference results.
   func runModel(onFrame pixelBuffer: CVPixelBuffer) -> Result? {
+    
     let sourcePixelFormat = CVPixelBufferGetPixelFormatType(pixelBuffer)
     assert(sourcePixelFormat == kCVPixelFormatType_32ARGB ||
              sourcePixelFormat == kCVPixelFormatType_32BGRA ||
@@ -222,43 +224,64 @@ class ModelDataHandler {
     isModelQuantized: Bool
   ) -> Data? {
     CVPixelBufferLockBaseAddress(buffer, .readOnly)
-    defer { CVPixelBufferUnlockBaseAddress(buffer, .readOnly) }
-    guard let mutableRawPointer = CVPixelBufferGetBaseAddress(buffer) else {
+    defer {
+      CVPixelBufferUnlockBaseAddress(buffer, .readOnly)
+    }
+    guard let sourceData = CVPixelBufferGetBaseAddress(buffer) else {
       return nil
     }
-    let count = CVPixelBufferGetDataSize(buffer)
-    let bufferData = Data(bytesNoCopy: mutableRawPointer, count: count, deallocator: .none)
-    var rgbBytes = [UInt8](repeating: 0, count: byteCount)
-    var index = 0
+    
+    let width = CVPixelBufferGetWidth(buffer)
+    let height = CVPixelBufferGetHeight(buffer)
+    let sourceBytesPerRow = CVPixelBufferGetBytesPerRow(buffer)
+    let destinationChannelCount = 3
+    let destinationBytesPerRow = destinationChannelCount * width
+    
+    var sourceBuffer = vImage_Buffer(data: sourceData,
+                                     height: vImagePixelCount(height),
+                                     width: vImagePixelCount(width),
+                                     rowBytes: sourceBytesPerRow)
+    
+    guard let destinationData = malloc(height * destinationBytesPerRow) else {
+      print("Error: out of memory")
+      return nil
+    }
+    
+    defer {
+        free(destinationData)
+    }
+
+    var destinationBuffer = vImage_Buffer(data: destinationData,
+                                          height: vImagePixelCount(height),
+                                          width: vImagePixelCount(width),
+                                          rowBytes: destinationBytesPerRow)
 
     let pixelBufferFormat = CVPixelBufferGetPixelFormatType(buffer)
-    var rgbChannelMap : [Int]
+
     switch (pixelBufferFormat) {
     case kCVPixelFormatType_32BGRA:
-      rgbChannelMap = [2, 1, 0]
-    case kCVPixelFormatType_32RGBA:
-      rgbChannelMap = [0, 1, 2]
-    case kCVPixelFormatType_32ABGR:
-      rgbChannelMap = [3, 2, 1]
+        vImageConvert_BGRA8888toRGB888(&sourceBuffer, &destinationBuffer, UInt32(kvImageNoFlags))
     case kCVPixelFormatType_32ARGB:
-      rgbChannelMap = [1, 2, 3]
+        vImageConvert_ARGB8888toRGB888(&sourceBuffer, &destinationBuffer, UInt32(kvImageNoFlags))
+    case kCVPixelFormatType_32RGBA:
+        vImageConvert_RGBA8888toRGB888(&sourceBuffer, &destinationBuffer, UInt32(kvImageNoFlags))
     default:
-      // Unknown pixel format.
-      return nil
+        // Unknown pixel format.
+        return nil
     }
 
-    // Iterate through pixels and reorder bytes to be in RGB order.
-    let numChannels = 4
-    for pixelIndex in 0..<count / numChannels {
-      let offset = pixelIndex * numChannels
-      for j in 0...2 {
-        rgbBytes[index] = bufferData[offset + rgbChannelMap[j]]
-        index += 1
-      }
+    let byteData = Data(bytes: destinationBuffer.data, count: destinationBuffer.rowBytes * height)
+    if isModelQuantized {
+        return byteData
     }
 
-    if isModelQuantized { return Data(bytes: rgbBytes) }
-    return Data(copyingBufferOf: rgbBytes.map { Float($0) / 255.0 })
+    // Not quantized, convert to floats
+    let bytes = Array<UInt8>(unsafeData: byteData)!
+    var floats = [Float]()
+    for i in 0..<bytes.count {
+        floats.append(Float(bytes[i]) / 255.0)
+    }
+    return Data(copyingBufferOf: floats)
   }
 }
 
