@@ -26,11 +26,7 @@ import sys
 
 import numpy as np
 from six.moves import xrange  # pylint: disable=redefined-builtin
-import tensorflow as tf
-
-from tensorflow.python.ops import io_ops
-from tensorflow.python.platform import gfile
-from tensorflow.python.util import compat
+import tensorflow.compat.v1 as tf
 
 from utils import tf_roll
 
@@ -57,7 +53,7 @@ def which_set(filename, validation_percentage, testing_percentage):
   base_name = os.path.basename(filename)
   hash_name = re.sub(r'_nohash_.*$', '', base_name)
 
-  hash_name_hashed = hashlib.sha1(compat.as_bytes(hash_name)).hexdigest()
+  hash_name_hashed = hashlib.sha1(tf.compat.as_bytes(hash_name)).hexdigest()
   percentage_hash = ((int(hash_name_hashed, 16) % (MAX_NUM_WAVS_PER_CLASS + 1))
                      * (100.0 / MAX_NUM_WAVS_PER_CLASS))
   if percentage_hash < validation_percentage:
@@ -73,7 +69,7 @@ def load_wav_file(filename):
   """Loads an audio file and returns a float PCM-encoded array of samples."""
   with tf.Session(graph=tf.Graph()) as sess:
     wav_filename_placeholder = tf.placeholder(tf.string, [])
-    wav_loader = io_ops.read_file(wav_filename_placeholder)
+    wav_loader = tf.io.read_file(wav_filename_placeholder)
     wav_decoder = tf.audio.decode_wav(wav_loader, desired_channels=1)
     return sess.run(
         wav_decoder, feed_dict={
@@ -89,7 +85,7 @@ def save_wav_file(filename, wav_data, sample_rate):
     wav_data_placeholder = tf.placeholder(tf.float32, [None, 1])
     wav_encoder = tf.audio.encode_wav(wav_data_placeholder,
                                       sample_rate_placeholder)
-    wav_saver = io_ops.write_file(wav_filename_placeholder, wav_encoder)
+    wav_saver = tf.io.write_file(wav_filename_placeholder, wav_encoder)
     sess.run(
         wav_saver,
         feed_dict={
@@ -131,7 +127,7 @@ class AudioProcessor(object):
   def prepare_data_index(self, silence_percentage, unknown_percentage,
                          wanted_words, validation_percentage,
                          testing_percentage):
-    """Prepares a list of the samples organized by set and label"""
+    """Prepares a list of the samples organized by set and label."""
     random.seed(RANDOM_SEED)
     wanted_words_index = {}
     for index, wanted_word in enumerate(wanted_words):
@@ -142,7 +138,7 @@ class AudioProcessor(object):
     # Look through all the subfolders to find audio samples
     for data_dir in self.data_dirs:
       search_path = os.path.join(data_dir, '*', '*.wav')
-      for wav_path in gfile.Glob(search_path):
+      for wav_path in tf.io.gfile.glob(search_path):
         word = re.search('.*/([^/]+)/.*.wav', wav_path).group(1).lower()
         # Treat the '_background_noise_' folder as a special case,
         # since we expect it to contain long audio samples we mix in
@@ -195,18 +191,18 @@ class AudioProcessor(object):
     self.word_to_index[SILENCE_LABEL] = SILENCE_INDEX
 
   def prepare_background_data(self):
-    """Searches a folder for background noise audio, and loads it into memory"""
+    """Searches a folder for background noise audio and loads it into memory."""
     self.background_data = []
     background_dir = os.path.join(self.data_dirs[0], BACKGROUND_NOISE_DIR_NAME)
     if not os.path.exists(background_dir):
       return self.background_data
     with tf.Session(graph=tf.Graph()) as sess:
       wav_filename_placeholder = tf.placeholder(tf.string, [])
-      wav_loader = io_ops.read_file(wav_filename_placeholder)
+      wav_loader = tf.io.read_file(wav_filename_placeholder)
       wav_decoder = tf.audio.decode_wav(wav_loader, desired_channels=1)
       search_path = os.path.join(self.data_dirs[0], BACKGROUND_NOISE_DIR_NAME,
                                  '*.wav')
-      for wav_path in gfile.Glob(search_path):
+      for wav_path in tf.io.gfile.glob(search_path):
         wav_data = sess.run(
             wav_decoder, feed_dict={
                 wav_filename_placeholder: wav_path
@@ -216,11 +212,11 @@ class AudioProcessor(object):
         raise Exception('No background wav files were found in ' + search_path)
 
   def prepare_processing_graph(self, model_settings):
-    """Builds a TensorFlow graph to apply the input distortions"""
+    """Builds a TensorFlow graph to apply the input distortions."""
     desired_samples = model_settings['desired_samples']
     self.wav_filename_placeholder_ = tf.placeholder(
         tf.string, [], name='filename')
-    wav_loader = io_ops.read_file(self.wav_filename_placeholder_)
+    wav_loader = tf.io.read_file(self.wav_filename_placeholder_)
     wav_decoder = tf.audio.decode_wav(
         wav_loader, desired_channels=1, desired_samples=desired_samples)
     # Allow the audio sample's volume to be adjusted.
@@ -268,7 +264,7 @@ class AudioProcessor(object):
                               model_settings['num_log_mel_features']]  # :13
 
   def set_size(self, mode):
-    """Calculates the number of samples in the dataset partition"""
+    """Calculates the number of samples in the dataset partition."""
     return len(self.data_index[mode])
 
   def get_data(self,
@@ -284,7 +280,7 @@ class AudioProcessor(object):
                sess,
                flip_frequency=0.0,
                silence_volume_range=0.0):
-    """Gather samples from the data set, applying transformations as needed"""
+    """Gather samples from the data set, applying transformations as needed."""
     # Pick one of the partitions to choose samples from.
     model_settings = self.model_settings
     candidates = self.data_index[mode]
@@ -406,7 +402,7 @@ class AudioProcessor(object):
     labels = []
     with tf.Session(graph=tf.Graph()) as sess:
       wav_filename_placeholder = tf.placeholder(tf.string, [], name='filename')
-      wav_loader = io_ops.read_file(wav_filename_placeholder)
+      wav_loader = tf.io.read_file(wav_filename_placeholder)
       wav_decoder = tf.audio.decode_wav(
           wav_loader, desired_channels=1, desired_samples=desired_samples)
       foreground_volume_placeholder = tf.placeholder(
@@ -430,7 +426,7 @@ class AudioProcessor(object):
     return data, labels
 
   def summary(self):
-    """Prints a summary of classes and label distributions"""
+    """Prints a summary of classes and label distributions."""
     set_counts = {}
     print('There are %d classes.' % (len(self.word_to_index)))
     print("1%% <-> %d samples in 'training'" % int(
