@@ -23,9 +23,8 @@ import os
 import time
 from absl import app
 from absl import flags
-import tensorflow as tf # TF2
+import tensorflow as tf
 import tensorflow_datasets as tfds
-assert tf.__version__.startswith('2')
 
 FLAGS = flags.FLAGS
 
