@@ -19,7 +19,7 @@ from __future__ import division
 from __future__ import print_function
 
 from absl import flags
-import tensorflow as tf # TF2
+import tensorflow as tf
 from tensorflow_examples.models.dcgan import dcgan
 
 FLAGS = flags.FLAGS
@@ -77,5 +77,4 @@ class DCGANBenchmark(tf.test.Benchmark):
     self.report_benchmark(wall_time=tf.reduce_mean(time_list[1:]))
 
 if __name__ == "__main__":
-  assert tf.__version__.startswith('2')
   tf.test.main()
