@@ -19,10 +19,9 @@ from __future__ import division
 from __future__ import print_function
 
 from absl import app
-import tensorflow as tf # TF2
+import tensorflow as tf
 from tensorflow_examples.models.densenet import densenet
 from tensorflow_examples.models.densenet import utils
-assert tf.__version__.startswith('2')
 
 
 class Train(object):
