@@ -19,7 +19,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import tensorflow as tf # TF2
+import tensorflow as tf
 
 
 class Encoder(tf.keras.Model):
