@@ -19,10 +19,10 @@ from __future__ import division
 from __future__ import print_function
 
 import time
-import tensorflow as tf # TF2
+
+import tensorflow as tf
 from tensorflow_examples.models.nmt_with_attention import train
 from tensorflow_examples.models.nmt_with_attention import utils
-assert tf.__version__.startswith('2')
 
 
 class NmtTest(tf.test.TestCase):
