@@ -20,10 +20,10 @@ from __future__ import division
 from __future__ import print_function
 
 from absl import app
-import tensorflow as tf # TF2
+
+import tensorflow as tf
 from tensorflow_examples.models.nmt_with_attention import nmt
 from tensorflow_examples.models.nmt_with_attention import utils
-assert tf.__version__.startswith('2')
 
 
 class Train(object):
