@@ -23,8 +23,8 @@ import os
 import time
 from absl import app
 from absl import flags
-import tensorflow as tf # TF2
-assert tf.__version__.startswith('2')
+
+import tensorflow as tf
 
 FLAGS = flags.FLAGS
 
